+++ packages/sdks/react-sdk/CHANGELOG.md
@@ -2,6 +2,30 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.26.5](https://github.com/descope/descope-js/compare/react-sdk-2.26.4...react-sdk-2.26.5) (2026-02-19)
+
+### Dependency Updates
+
+* `user-management-widget` updated to version `0.11.10`
+* `user-profile-widget` updated to version `0.9.5`
+## [2.26.4](https://github.com/descope/descope-js/compare/react-sdk-2.26.3...react-sdk-2.26.4) (2026-02-17)
+
+### Dependency Updates
+
+* `user-management-widget` updated to version `0.11.9`
+* `user-profile-widget` updated to version `0.9.4`
+* `outbound-applications-widget` updated to version `0.2.21`
+* `web-component` updated to version `3.55.3`
+* `tenant-profile-widget` updated to version `0.5.21`
+## [2.26.3](https://github.com/descope/descope-js/compare/react-sdk-2.26.2...react-sdk-2.26.3) (2026-02-11)
+
+### Dependency Updates
+
+* `user-management-widget` updated to version `0.11.8`
+* `user-profile-widget` updated to version `0.9.3`
+* `outbound-applications-widget` updated to version `0.2.20`
+* `web-component` updated to version `3.55.2`
+* `tenant-profile-widget` updated to version `0.5.20`
 ## [2.26.2](https://github.com/descope/descope-js/compare/react-sdk-2.26.1...react-sdk-2.26.2) (2026-02-01)
 
 ### Dependency Updates